+++ doc/libxorif_html/search/variables_70.js
@@ -1,4 +1,5 @@
 var searchData=
 [
+  ['prach_5fss',['prach_ss',['../structxorif__stall__monitor.html#ade8c5e8c14408a13b3c145560b2eb3ac',1,'xorif_stall_monitor']]],
   ['prime_5fcc',['prime_cc',['../structxocp__state__t.html#a90439d304c97612eb31c7fc3ed555e68',1,'xocp_state_t']]]
 ];