+++ doc/libxorif_html/search/variables_75.js
@@ -7,5 +7,7 @@ var searchData=
   ['ul_5fctrl_5fsize',['ul_ctrl_size',['../structxorif__cc__alloc.html#a2c058335d1363535e7fac34fd462d321',1,'xorif_cc_alloc']]],
   ['ul_5fctrl_5fsym_5fnum',['ul_ctrl_sym_num',['../structxorif__cc__alloc.html#ab3369e2390d6b04d52fb38c56163e965',1,'xorif_cc_alloc']]],
   ['ul_5fradio_5fch_5fdly',['ul_radio_ch_dly',['../structxorif__cc__config.html#af9b342510d0c4d7681875f3023294b0f',1,'xorif_cc_config']]],
-  ['ul_5fupdate',['ul_update',['../structxocp__triggers.html#ab4bf1e198fc20d416e3d31fd3bcf1978',1,'xocp_triggers']]]
+  ['ul_5fss',['ul_ss',['../structxorif__stall__monitor.html#a0a172edc4b7c8d342518c9ae9c07d5dd',1,'xorif_stall_monitor']]],
+  ['ul_5fupdate',['ul_update',['../structxocp__triggers.html#ab4bf1e198fc20d416e3d31fd3bcf1978',1,'xocp_triggers']]],
+  ['unsol_5fss',['unsol_ss',['../structxorif__stall__monitor.html#a7fd848f262579ce09eabd4b94f0f59b1',1,'xorif_stall_monitor']]]
 ];