+++ scripts/runvp.bash
@@ -1,8 +1,20 @@
 #!/bin/bash 
-#source /proj/petalinux/2020.2/petalinux-v2020.2_daily_latest/tool/petalinux-v2020.2-final/settings.csh
+
+cat ./runvp_help.bash
 
 ## This is a global
-ipRepo="-tclargs /proj/ipeng_scratch/maginnis/w/IP3_maginnis_v30l/DEV/output/internal/vivado/data/ip/xilinx/oran_radio_if_v3_0/"
+## Now pick up Env Variable. Set doing sometrhin like this.
+#setenv XORIF_REPO "/proj/ipeng_scratch/maginnis/w/IP3_maginnis_v30l/DEV/output/internal/vivado/data/ip/xilinx/oran_radio_if_v3_0/"
+#ipRepo="-tclargs /proj/ipeng_scratch/maginnis/w/IP3_maginnis_v30l/DEV/output/internal/vivado/data/ip/xilinx/oran_radio_if_v3_0/"
+if [ -z $XORIF_REPO ]
+then
+  ipRepo=""
+  echo "XORIF_REPO is Empty, using Release Repo: $XORIF_REPO\n$ipRepo"
+else
+  ipRepo="-tclargs $XORIF_REPO"
+  echo "XORIF_REPO is Set, using repo: $XORIF_REPO"
+  echo "TCLARGS set to $ipRepo"
+fi
 
 ## This is a serial script. Run over night. If the config list gets too big, parallelise
 
@@ -63,8 +75,9 @@ function runVpBuild {
   ipMode=$2
   exMode=$3
   toolVersion=$4
+  extras=$5
 
-  runVivadoBuild ${board} ${ipMode} ${exMode}
+  runVivadoBuild ${board} ${ipMode} ${exMode} ${extras}
   buildName=$globalString
   echo "Running Petalinux Build for $buildName"
   runPetalinuxBuild $buildName $toolVersion
@@ -81,13 +94,23 @@ doPetaLinux=1
 ##------------------------------------------------------------------------------
 ## Builds to run, serially.
 ##------------------------------------------------------------------------------
-#runVpBuild zcu111 e1x9000x25_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144 Exd 2022_2 ""
-#runVpBuild zcu111 e1x9000x25_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144 Exs 2022_2 "Loop"
-runVpBuild zcu111 e1x9000x25_ss8x4_cc4x6600_dcc16x4x4x16x16_aptw262144        Exd 2022_2 ""
-#runVpBuild zcu111 e1x9000x25_ss8x4_cc4x6600_dcc16x4x4x16x16_aptw262144        Exs 2022_2 "Loop"
-#runVpBuild zcu111 e1x9000x10_ss8x8_cc4x6600_dcc12x4x4x16x8_ocpi               Exd 2022_2 ""
-#runVpBuild zcu111 e1x9000x10_ss8x8_cc4x6600_dcc12x4x4x16x8_ocpi               Exs 2022_2 "Loop"
-#runVpBuild zcu111 e4x8000x25_ss4x4_cc4x6600_dcc16x4x4x12x8_swtest             Exd 2022_2 ""
-#runVpBuild zcu111 e4x8000x10_ss4x4_cc4x6600_dcc16x4x4x12x8_swtest             Exd 2022_2 ""
+#runVpBuild zcu111 e1x9600x25_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144       Exd 2023_2 ""
+#runVpBuild zcu111 e1x9600x25_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144       Exs 2023_2 "Loop"
+#runVpBuild zcu111 e1x9600x25_ss8x4_cc4x6600_dcc16x4x4x16x16_aptw262144              Exd 2023_2 ""
+#runVpBuild zcu111 e1x9600x25_ss8x4_cc4x6600_dcc16x4x4x16x16_aptw262144              Exs 2023_2 "Loop"
+#runVpBuild zcu111 e1x9000x10_ss8x8_cc4x6600_dcc12x4x4x16x8_ocpi                     Exd 2023_2 ""
+#runVpBuild zcu670 e1x9600x25_ss8x4_cc4x6600_dcc16x4x4x16x16_aptw262144              Exs 2023_2 "Loop"
+#runVpBuild zcu111 e1x9000x10_ss8x8_cc4x6600_dcc12x4x4x16x8_ocpi                     Exs 2023_2 "Loop"
+#runVpBuild zcu111 e4x8000x25_ss4x4_cc4x6600_dcc16x4x4x12x8_swtest                   Exd 2023_2 ""
+#runVpBuild zcu111 e4x8000x10_ss4x4_cc4x6600_dcc16x4x4x12x8_swtest                   Exd 2023_2 ""
+
+#runVpBuild zcu111 e1x9600x10_ss8x8_cc4x6600_dcc12x4x4x16x8_cmw00i_opt               Exd 2023_2 ""
+#runVpBuild zcu111 e1x9600x10_ss8x8_cc4x6600_dcc12x4x4x16x8_ocpi_opt                 Exd 2023_2 ""
+#runVpBuild zcu111 e1x9600x10_ss8x8_cc4x6600_dcc12x4x4x16x8_cmw00i_ocpi_opt          Exd 2023_2 ""
+
+###runVpBuild zcu111 e1x9600x10_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144       Exd 2023_2 ""
+#runVpBuild zcu111 e1x9600x10_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144map11  Exd 2023_2 ""
+
+