+++ scripts/runvp_slv.bash
@@ -1,0 +1,3 @@
+#!/bin/bash 
+source ./runvp.bash
+runVpBuild zcu111 e1x9600x10_ss8x4_cc4x6600_dcc16x4x4x16x16_cmw00i_aptw262144       Exd 2023_2 ""